requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cmake==3.25.2
colorama==0.4.6
comtypes==1.4.5
contourpy==1.2.1
cycler==0.12.1
deepface==0.0.92
dlib==19.24.2
easyocr==1.7.1
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
gunicorn==22.0.0
h5py==3.11.0
httplib2==0.22.0
idna==3.7
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.8
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pypiwin32==223
pyserial==3.5
PySocks==1.7.1
python-bidi==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.32.3
retina-face==0.0.17
rich==13.7.1
rsa==4.9
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
shapely==2.0.5
six==1.16.0
soupsieve==2.5
SpeechRecognition==3.10.4
sympy==1.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.17.0
tifffile==2024.7.24
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.64
ultralytics-thop==2.0.0
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
